Move table_exports.py from dags/config to include
